packages/pykit-observability/pyproject.toml
[project]
name = "pykit-observability"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "pykit-errors",
    "pykit-component",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pykit_observability"]
